Refactor booking service and controller for DTO support
